refactor, downloadable midi on output
