+++ dashboard/streamlit/prompts.py
@@ -12,7 +12,7 @@ from matplotlib import pyplot as plt
 
 from data.quantizer import MidiQuantizer
 from utils import piece_av_files, decode_and_output
-from dashboard.components import piece_selector, download_button
+from dashboard.streamlit.components import piece_selector, download_button
 from data.dataset import MyTokenizedMidiDataset, quantized_piece_to_records
 from data.tokenizer import MidiEncoder, VelocityEncoder, QuantizedMidiEncoder
